Move components outside of the routes
Components should be reusable and should not belong to any route
Part of #125
