+++ reposilite-frontend/src/routes/dashboard/DashboardView.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script>
-    import Navbar from './components/Navbar.vue';
+    import Navbar from '../../components/Navbar.vue';
 
     export default {
         name: 'DashboardView',
