+++ reposilite-frontend/src/routes/home/HomeView.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-    import Search from './components/Search.vue';
+    import Search from '../../components/Search.vue';
 
     export default {
         name: 'HomeView',
